Finished Search component and API call
